feat(Config): swap Config over TS
